use internal node names for block device actions (#23)
qemu has multiple ways to attach disks (-hda, -drive, -blockdev), all
share in common that the devices attached have device and node names.

Using the device names for the operations is not really ideal, so switch
to the internal node-name instead. This is also required for future
implementation of copy-before-write filters to allow for image during
fleecing during backup.

This change may break incemental backups for existing directories.
Before the target files were written to:

 dir/ide0-hd0/..

where ide0-hd0 was the device name. Now the real node name for the disk
is used, (-drive node-name=disk0,file=..), so the target directory is:

 dir/disk0/..